requirements.txt
intel-extension-for-pytorch==2.1.20
pytorch==2.1.0
torchvision==0.16.0